backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
opencv-python
numpy
face_recognition
python-dotenv
reportlab
